heuristic data type fixed

+++ src/code/Node.java
@@ -22,11 +22,11 @@ public class Node<State, Operator> implements Comparable<Node<State, Operator>> 
     private Node _parent;
     private Operator _action;
     private int[] _pathCost;
-    private float _heuristic;
+    private int _heuristic;
     private int _depth;
     private boolean _isgreedy;
 
-    public Node(State state, Node parent, Operator action, int[] pathCost, float heuristic, int depth) {
+    public Node(State state, Node parent, Operator action, int[] pathCost, int heuristic, int depth) {
         _state = state;
         _parent = parent;
         _action = action;
@@ -46,7 +46,7 @@ public class Node<State, Operator> implements Comparable<Node<State, Operator>> 
         _parent = null;
         _action = null;
         _pathCost = null;
-        _heuristic = 0.0f;
+        _heuristic = 0;
         _depth = 0;
     }
 
@@ -91,7 +91,7 @@ public class Node<State, Operator> implements Comparable<Node<State, Operator>> 
         return _heuristic;
     }
 
-    public void setHeuristic(float heuristic) {
+    public void setHeuristic(int heuristic) {
         _heuristic = heuristic;
     }
